requirements.txt
crewai
Langchain_google_genai
load_dotenv
crewai_tools
